# PetsClient/pets-swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1",
    "title": "Pets API"
  },
  "paths": {
    "/api/Pets": {
      "get": {
        "tags": [ "Pets" ],
        "operationId": "ApiPetsGet",
        "consumes": [],
        "produces": [ "text/plain", "application/json", "text/json" ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "uniqueItems": false,
              "type": "array",
              "items": { "$ref": "#/definitions/Pet" }
            }
          }
        }
      },
      "post": {
        "tags": [ "Pets" ],
        "operationId": "ApiPetsPost",
        "consumes": [ "application/json-patch+json", "application/json", "text/json", "application/*+json" ],
        "produces": [ "text/plain", "application/json", "text/json" ],
        "parameters": [
          {
            "name": "pet",
            "in": "body",
            "required": false,
            "schema": { "$ref": "#/definitions/Pet" }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/Pet" }
          },
          "400": {
            "description": "Bad Request",
            "schema": { "$ref": "#/definitions/ProblemDetails" }
          },
          "0": { "schema": { "$ref": "#/definitions/ProblemDetails" } }
        }
      }
    },
    "/api/Pets/{id}": {
      "get": {
        "tags": [ "Pets" ],
        "operationId": "ApiPetsByIdGet",
        "consumes": [],
        "produces": [ "text/plain", "application/json", "text/json" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/Pet" }
          },
          "404": {
            "description": "Not Found",
            "schema": { "$ref": "#/definitions/ProblemDetails" }
          },
          "0": { "schema": { "$ref": "#/definitions/ProblemDetails" } }
        }
      },
      "put": {
        "tags": [ "Pets" ],
        "operationId": "ApiPetsByIdPut",
        "consumes": [ "application/json-patch+json", "application/json", "text/json", "application/*+json" ],
        "produces": [ "text/plain", "application/json", "text/json" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pet",
            "in": "body",
            "required": false,
            "schema": { "$ref": "#/definitions/Pet" }
          }
        ],
        "responses": {
          "204": { "description": "Success" },
          "404": {
            "description": "Not Found",
            "schema": { "$ref": "#/definitions/ProblemDetails" }
          },
          "400": {
            "description": "Bad Request",
            "schema": { "$ref": "#/definitions/ProblemDetails" }
          },
          "0": { "schema": { "$ref": "#/definitions/ProblemDetails" } }
        }
      },
      "delete": {
        "tags": [ "Pets" ],
        "operationId": "ApiPetsByIdDelete",
        "consumes": [],
        "produces": [ "text/plain", "application/json", "text/json" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/Pet" }
          },
          "404": {
            "description": "Not Found",
            "schema": { "$ref": "#/definitions/ProblemDetails" }
          },
          "400": {
            "description": "Bad Request",
            "schema": { "$ref": "#/definitions/ProblemDetails" }
          },
          "0": { "schema": { "$ref": "#/definitions/ProblemDetails" } }
        }
      }
    }
  },
  "definitions": {
    "Pet": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "owner": { "type": "string" },
        "name": { "type": "string" },
        "type": { "type": "string" },
        "age": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "ProblemDetails": {
      "type": "object",
      "properties": {
        "type": { "type": "string" },
        "title": { "type": "string" },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "detail": { "type": "string" },
        "instance": { "type": "string" }
      },
      "additionalProperties": { "type": "object" }
    }
  }
}